Fixing how I insert user data to DB on signup and simplifying App component code
