+++ src/App.js
@@ -8,7 +8,7 @@ import UpdateProfile from './Routes/UpdateProfile';
 import AddNewLyrics from './Routes/AddNewLyrics';
 import NewLyricSearch from './Routes/NewLyricSearch';
 import EditFavoriteLyric from './Routes/EditFavoriteLyric';
-import { updateAll, updateEmail, updateFirstName, updateIsLoggedIn, updateLastName } from './Utilities/Redux/userSlice';
+import { updateEmail, updateFirstName, updateId, updateIsLoggedIn, updateLastName } from './Utilities/Redux/userSlice';
 import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useSquid, useCollection } from '@squidcloud/react'
@@ -56,21 +56,7 @@ function App() {
           console.log('User is signed in: ', user);
           const uid = user.uid;
 
-          //Split out first/last names
-          const fullName = user.displayName;
-
-          // TO DO: GATHER NAME INFO ON SIGN UP
-          // const [firstName, ...lastNames] = fullName.split(' ');
-          // const lastName = lastNames.join(' ');
-  
-          // console.log('first name: ', firstName);
-          // console.log('last name: ', lastName);
-          dispatch(updateAll({
-            id: uid,
-            email: user.email,
-            firstName: fullName,
-            lastName: fullName,
-          }))
+          dispatch(updateId(uid));
         
         } else {
           console.log('No user is signed in');
@@ -80,52 +66,22 @@ function App() {
   
     const usersCollection = useCollection('users', 'postgres_id'); // Reference to users collection in DB
   
-    // ***** TO DO EVENTUALLY ***** -- update to only have the user information query/update once after login instead of every time page loads
-    // ^^ will setting up the session stuff fix that?
-  
       // Store initial user info in variables to use here
       const userInformation = useSelector(state => state.user.userInfo);
-      const { id, email, firstName, lastName } = userInformation;
+      const { id } = userInformation;
     
   
-      // Check to see if current user exists in DB - triggers other functions to update Redux, insert new users into DB
+      // Get current user info from Postgres
       useEffect(() => {
         if (user) {
           console.log('doop');
           (async () => {
-            const doesUserExist = await userExists();
-            console.log('does the user exist? ', doesUserExist);
-            
-            if (doesUserExist) {
-              // if user exists in DB, update state with the stored information
-              getUserInfo();
-            }
-            if (!doesUserExist) {
-                console.log('User does not exist - about to run insertUser');
-              // if user doesn't exist in DB yet, insert them into the DB
-              insertNewUser();
-            }
+            getUserInfo();
           })();
         }
       // eslint-disable-next-line react-hooks/exhaustive-deps
       }, [user]);
   
-  
-      // Function to check if user file exists in the DB
-      const userExists = async () => {
-        const userDoc = await usersCollection
-        .doc({ id: id })
-        .snapshot();
-        console.log('userQuery during userExists: ', userDoc);
-        if (userDoc) {
-          // if the user is in the DB, return true
-          return true;
-        }
-        else {
-          // if the user is not in the DB, return false
-          return false;
-        }
-    };
   
       // Function to query Postgres for user data and update user info stored in Redux
       const getUserInfo = async () => {
@@ -149,20 +105,7 @@ function App() {
         }     
       };
   
-  
-  // Function to insert data into Postgres for new user
-  const insertNewUser = async () => {
-    console.log("Running inserNewUser");
-    console.log("Form data being inserted: ", userInformation);
-    await usersCollection.doc({ id: id }).insert({
-        id: id,
-        first_name: firstName,
-        last_name: lastName,
-        email: email,
-    })
-        .then(() => console.log("New user inserted into DB successfully"))
-        .catch((err) => console.error("Error inserting new user into DB: ", err));
-  }
+
 // Routing
 const router = createBrowserRouter([
   {
