+++ src/Routes/Signup.jsx
@@ -2,6 +2,7 @@ import React, { useRef, useState, useEffect } from 'react'
 import { useNavigate, Link } from 'react-router-dom';
 import { Alert, Container, Card, Form, Button } from 'react-bootstrap';
 import { getAuth, createUserWithEmailAndPassword } from 'firebase/auth';
+import { useCollection } from '@squidcloud/react';
 
 export default function Signup() {
     const emailRef = useRef();
@@ -14,10 +15,11 @@ export default function Signup() {
 
     const [error, setError] = useState('');
     const [loading, setLoading] = useState(false);
-      const [isLoggedIn, setIsLoggedIn] = useState(false);
+    const [isLoggedIn, setIsLoggedIn] = useState(false);
+
+    const usersCollection = useCollection('users', 'postgres_id'); // Reference to users collection in DB
 
     // TO DO: ADD ERROR FOR EMAIL EXISTS
-    // TO DO: WRITE FUNCTION TO ADD USER F/L NAMES TO DB
 
     function handleSubmit(e) {
         e.preventDefault();
@@ -30,6 +32,19 @@ export default function Signup() {
 
         createUserWithEmailAndPassword(auth, emailRef.current.value, passwordRef.current.value)
             .then((userCredential) => {
+                const insertNewUser = async () => {
+                    console.log("Running inserNewUser");
+                    const id = userCredential.user.uid;
+                    await usersCollection.doc({ id: id }).insert({
+                        id: id,
+                        first_name: firstNameRef.current.value,
+                        last_name: lastNameRef.current.value,
+                        email: emailRef.current.value,
+                    })
+                        .then(() => console.log("New user inserted into DB successfully"))
+                        .catch((err) => console.error("Error inserting new user into DB: ", err));
+                  }
+                insertNewUser();
                 // Signed up
                 // const user = userCredential.user; <-- do i need that for anything? maybe can run a function here to add user to DB
                 // navigate('/dashboard');
